+++ src/Resources/app/administration/src/module/swag-paypal-settings/view/swag-paypal-settings-storefront/index.ts
@@ -51,6 +51,25 @@ export default Shopware.Component.wrapComponentConfig({
                 && !this.settingsStore.getActual('SwagPayPal.settings.ecsListingEnabled');
         },
 
+        ecsShippingCallbackSettingDisabled(): boolean {
+            return this.ecsSettingsDisabled || !!this.settingsStore.getActual('SwagPayPal.settings.isLocalEnvironment');
+        },
+
+        ecsShippingCallbackSettingTooltip(): { message: string; showOnDisabled?: boolean; disabled?: boolean } {
+            if (this.settingsStore.getActual('SwagPayPal.settings.isLocalEnvironment')) {
+                return {
+                    message: this.$t('swag-paypal-settings.express.ecsShippingCallbackSettingDisabledTooltip'),
+                    showOnDisabled: true,
+                    disabled: false,
+                };
+            }
+
+            return {
+                message: '',
+                disabled: true,
+            };
+        },
+
         systemConfigRepository() {
             return this.repositoryFactory.create('system_config');
         },